Restructure packaging for await component
COAND-497
