+++ action/src/main/java/com/adyen/checkout/action/ActionDelegateProvider.kt
@@ -12,8 +12,8 @@ import android.app.Application
 import androidx.lifecycle.SavedStateHandle
 import com.adyen.checkout.adyen3ds2.Adyen3DS2Configuration
 import com.adyen.checkout.adyen3ds2.internal.provider.Adyen3DS2ComponentProvider
-import com.adyen.checkout.await.AwaitComponentProvider
 import com.adyen.checkout.await.AwaitConfiguration
+import com.adyen.checkout.await.internal.provider.AwaitComponentProvider
 import com.adyen.checkout.components.base.ActionDelegate
 import com.adyen.checkout.components.base.BaseConfigurationBuilder
 import com.adyen.checkout.components.base.ComponentParams
