+++ await/src/test/java/com/adyen/checkout/await/AwaitComponentTest.kt
@@ -12,6 +12,8 @@ import android.app.Activity
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.viewModelScope
 import app.cash.turbine.test
+import com.adyen.checkout.await.internal.ui.AwaitComponentViewType
+import com.adyen.checkout.await.internal.ui.AwaitDelegate
 import com.adyen.checkout.components.ActionComponentEvent
 import com.adyen.checkout.components.base.ActionComponentEventHandler
 import com.adyen.checkout.components.model.payments.response.AwaitAction
